requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
alembic==1.15.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.3.0
bcrypt==4.0.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
comm==0.2.2
cryptography==45.0.2
dataclasses-json==0.6.7
debugpy==1.8.13
decorator==5.2.1
distro==1.9.0
docker==7.1.0
exceptiongroup==1.2.2
executing==2.2.0
faiss-cpu==1.7.4
fastapi==0.115.8
filelock==3.18.0
filetype==1.2.0
frozenlist==1.5.0
fsspec==2025.5.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
ipinfo==5.1.1
ipykernel==6.29.5
ipython==8.34.0
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.15
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.46
langchain-google-genai==2.1.0
langchain-text-splitters==0.3.7
langsmith==0.3.18
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.2.0
mypy-extensions==1.0.0
mysql-connector==2.2.9
narwhals==1.31.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parso==0.8.4
passlib==1.7.4
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.3
psutil==7.0.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.7.0
PyMySQL==1.1.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
pywin32==310
PyYAML==6.0.2
pyzmq==26.3.0
referencing==0.36.2
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.23.1
rsa==4.9
secure==0.3.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.39
SQLAlchemy-Utils==0.41.2
sqlmodel==0.0.18
stack-data==0.6.3
starlette==0.45.3
streamlit==1.43.2
sympy==1.14.0
tenacity==9.0.0
tldextract==5.1.3
toml==0.10.2
torch==2.7.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.2
watchdog==6.0.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
widgetsnbextension==4.0.14
yarl==1.18.3
zenml==0.82.1
zstandard==0.23.0
